+++ xrview/core/__init__.py
@@ -5,9 +5,9 @@ from xrview.core.plot import BasePlot
 from xrview.core.viewer import BaseViewer
 
 __all__ = [
-    'BasePanel',
-    'GridPlot',
-    'SpacerPanel',
-    'BasePlot',
-    'BaseViewer',
+    "BasePanel",
+    "GridPlot",
+    "SpacerPanel",
+    "BasePlot",
+    "BaseViewer",
 ]