+++ xrview/core/plot.py
@@ -8,9 +8,9 @@ from xrview.core.panel import BasePanel
 from xrview.elements import Element
 from xrview.glyphs import get_glyph_list
 from xrview.handlers import DataHandler
-from xrview.mappers import map_figures_and_glyphs, _get_overlay_figures
+from xrview.mappers import _get_overlay_figures, map_figures_and_glyphs
 from xrview.palettes import RGB
-from xrview.utils import is_dataarray, is_dataset, clone_models, rsetattr
+from xrview.utils import clone_models, is_dataarray, is_dataset, rsetattr
 
 
 class BasePlot(BasePanel):
@@ -18,23 +18,27 @@ class BasePlot(BasePanel):
 
     element_type = Element
     handler_type = DataHandler
-    default_tools = 'pan,wheel_zoom,save,reset,'
+    default_tools = "pan,wheel_zoom,save,reset,"
 
-    def __init__(self, data, x,
-                 overlay='dims',
-                 coords=None,
-                 glyphs='line',
-                 title=None,
-                 share_y=False,
-                 tooltips=None,
-                 tools=None,
-                 toolbar_location='right',
-                 figsize=(600, 300),
-                 ncols=1,
-                 palette=None,
-                 ignore_index=False,
-                 theme=None,
-                 **fig_kwargs):
+    def __init__(
+        self,
+        data,
+        x,
+        overlay="dims",
+        coords=None,
+        glyphs="line",
+        title=None,
+        share_y=False,
+        tooltips=None,
+        tools=None,
+        toolbar_location="right",
+        figsize=(600, 300),
+        ncols=1,
+        palette=None,
+        ignore_index=False,
+        theme=None,
+        **fig_kwargs,
+    ):
         """ Constructor.
 
         Parameters
@@ -80,23 +84,24 @@ class BasePlot(BasePanel):
         # check data
         if is_dataarray(data):
             if data.name is None:
-                self.data = data.to_dataset(name='Data')
+                self.data = data.to_dataset(name="Data")
             else:
                 self.data = data.to_dataset()
         elif is_dataset(data):
             self.data = data
         else:
-            raise ValueError('data must be xarray DataArray or Dataset.')
+            raise ValueError("data must be xarray DataArray or Dataset.")
 
         # check x
         if x in self.data.dims:
             self.x = x
         else:
             raise ValueError(
-                x + ' is not a dimension of the provided dataset.')
+                x + " is not a dimension of the provided dataset."
+            )
 
         # check overlay
-        if overlay in ('dims', 'data_vars'):
+        if overlay in ("dims", "data_vars"):
             self.overlay = overlay
         else:
             raise ValueError('overlay must be "dims" or "data_vars"')
@@ -123,7 +128,7 @@ class BasePlot(BasePanel):
         if tools is None:
             self.tools = self.default_tools
             if self.tooltips is not None:
-                self.tools += 'hover,'
+                self.tools += "hover,"
         else:
             self.tools = tools
 
@@ -137,21 +142,25 @@ class BasePlot(BasePanel):
         plot_data = dict()
 
         if coords is True:
-            coords = [c for c in data.coords
-                      if self.x in data[c].dims and c != self.x]
+            coords = [
+                c
+                for c in data.coords
+                if self.x in data[c].dims and c != self.x
+            ]
 
         for v in list(data.data_vars) + (coords or []):
             if self.x not in data[v].dims:
-                raise ValueError(self.x + ' is not a dimension of ' + v)
+                raise ValueError(self.x + " is not a dimension of " + v)
             elif len(data[v].dims) == 1:
                 plot_data[v] = data[v].values
             elif len(data[v].dims) == 2:
                 dim = [d for d in data[v].dims if d != self.x][0]
                 for d in data[dim].values:
-                    plot_data[v + '_' + str(d)] = \
+                    plot_data[v + "_" + str(d)] = (
                         data[v].sel(**{dim: d}).values
+                    )
             else:
-                raise ValueError(v + ' has too many dimensions')
+                raise ValueError(v + " has too many dimensions")
 
         # TODO: doesn't work for irregularly sampled data
         if self.ignore_index:
@@ -161,13 +170,16 @@ class BasePlot(BasePanel):
                 else:
                     freq = data.indexes[self.x].freq
                 index = pd.DatetimeIndex(
-                    start=0, freq=freq, periods=data.sizes[self.x])
+                    start=0, freq=freq, periods=data.sizes[self.x]
+                )
             else:
                 index = np.arange(data.sizes[self.x])
         else:
             if isinstance(data.indexes[self.x], pd.MultiIndex):
-                index = [tuple(str(i) for i in idx)
-                         for idx in self.data.indexes[self.x].tolist()]
+                index = [
+                    tuple(str(i) for i in idx)
+                    for idx in self.data.indexes[self.x].tolist()
+                ]
                 for n in data.indexes[self.x].names:
                     plot_data[n] = data.indexes[self.x].get_level_values(n)
             else:
@@ -197,9 +209,18 @@ class BasePlot(BasePanel):
     def _make_maps(self):
         """ Make the figure and glyph map. """
         self.figure_map, self.glyph_map = map_figures_and_glyphs(
-            self.data, self.x, self.handlers, self.glyphs, self.overlay,
-            self.fig_kwargs, self.added_figures, self.added_overlays,
-            self.added_overlay_figures, self.palette, self.title)
+            self.data,
+            self.x,
+            self.handlers,
+            self.glyphs,
+            self.overlay,
+            self.fig_kwargs,
+            self.added_figures,
+            self.added_overlays,
+            self.added_overlay_figures,
+            self.palette,
+            self.title,
+        )
 
     def _make_figures(self):
         """ Make figures. """
@@ -209,23 +230,28 @@ class BasePlot(BasePanel):
         for _, f in self.figure_map.iterrows():
             # adjust x axis type for datetime x values
             if isinstance(self.data.indexes[self.x], pd.DatetimeIndex):
-                f.fig_kwargs['x_axis_type'] = 'datetime'
+                f.fig_kwargs["x_axis_type"] = "datetime"
 
             # set axis ranges
             if len(self.figures) == 0:
                 if isinstance(self.data.indexes[self.x], pd.MultiIndex):
-                    f.fig_kwargs['x_range'] = FactorRange(
-                        *(tuple(str(i) for i in idx)
-                          for idx in self.data.indexes[self.x].tolist()),
-                        range_padding=0.1)
+                    f.fig_kwargs["x_range"] = FactorRange(
+                        *(
+                            tuple(str(i) for i in idx)
+                            for idx in self.data.indexes[self.x].tolist()
+                        ),
+                        range_padding=0.1,
+                    )
             else:
-                f.fig_kwargs['x_range'] = self.figures[0].x_range
+                f.fig_kwargs["x_range"] = self.figures[0].x_range
                 if self.share_y:
-                    f.fig_kwargs['y_range'] = self.figures[0].y_range
+                    f.fig_kwargs["y_range"] = self.figures[0].y_range
 
             if self.figsize is not None:
-                width = self.figsize[0]//self.ncols
-                height = self.figsize[1]//self.figure_map.shape[0]*self.ncols
+                width = self.figsize[0] // self.ncols
+                height = (
+                    self.figsize[1] // self.figure_map.shape[0] * self.ncols
+                )
                 f.fig_kwargs.update(dict(plot_width=width, plot_height=height))
 
             self.figures.append(figure(tools=self.tools, **f.fig_kwargs))
@@ -236,21 +262,26 @@ class BasePlot(BasePanel):
             glyph_kwargs = clone_models(g.glyph_kwargs)
             if isinstance(g.method, str):
                 getattr(self.figures[g.figure], g.method)(
-                    source=g.handler.source, **glyph_kwargs)
+                    source=g.handler.source, **glyph_kwargs
+                )
             else:
                 self.figures[g.figure].add_layout(
-                    g.method(source=g.handler.source, ** glyph_kwargs))
+                    g.method(source=g.handler.source, **glyph_kwargs)
+                )
             # add an invisible circle glyph to make glyph selectable
-            if g.method != 'circle':
+            if g.method != "circle":
                 self.figures[g.figure].circle(
-                    source=g.handler.source, size=0,
-                    **{'x': glyph_kwargs[g.x_arg], 'y': glyph_kwargs[g.y_arg]})
+                    source=g.handler.source,
+                    size=0,
+                    **{"x": glyph_kwargs[g.x_arg], "y": glyph_kwargs[g.y_arg]},
+                )
 
     def _add_annotations(self):
         """ Add annotations. """
         for idx, a in enumerate(self.added_annotations):
             f_idx = _get_overlay_figures(
-                self.added_annotation_figures[idx], self.figure_map)
+                self.added_annotation_figures[idx], self.figure_map
+            )
             for f in f_idx:
                 if isinstance(a, Glyph):
                     self.figures[f].add_glyph(a)
@@ -264,12 +295,15 @@ class BasePlot(BasePanel):
             for f in self.figures:
                 f.select(HoverTool).tooltips = tooltips
                 if isinstance(self.data.indexes[self.x], pd.DatetimeIndex):
-                    f.select(HoverTool).formatters = {'index': 'datetime'}
+                    f.select(HoverTool).formatters = {"index": "datetime"}
 
     def _finalize_layout(self):
         """ Finalize layout. """
-        self.layout = gridplot(self.figures, ncols=self.ncols,
-                               toolbar_location=self.toolbar_location)
+        self.layout = gridplot(
+            self.figures,
+            ncols=self.ncols,
+            toolbar_location=self.toolbar_location,
+        )
 
     def _modify_figures(self):
         """ Modify the attributes of multiple figures. """
@@ -303,7 +337,7 @@ class BasePlot(BasePanel):
 
         return self.layout
 
-    def add_figure(self, data, glyphs='line', coords=None, name=None):
+    def add_figure(self, data, glyphs="line", coords=None, name=None):
         """ Add a figure to the layout.
 
         Parameters
@@ -325,8 +359,9 @@ class BasePlot(BasePanel):
         element = self.element_type(glyphs, data, coords, name)
         self.added_figures.append(element)
 
-    def add_overlay(self, data, glyphs='line', coords=None, name=None,
-                    onto=None):
+    def add_overlay(
+        self, data, glyphs="line", coords=None, name=None, onto=None
+    ):
         """ Add an overlay to a figure in the layout.
 
         Parameters