+++ xrview/glyphs.py
@@ -1,20 +1,38 @@
 """ xrview.glyphs """
+from bokeh.models import Band as _Band
+from bokeh.models import Whisker as _Whisker
+
 from xrview.utils import MappingProxyType
 
-from bokeh.models import Whisker as _Whisker, Band as _Band
-
-__all__ = ['Line', 'Circle', 'Diamond', 'Square', 'Triangle', 'Ray', 'HBar',
-           'VBar', 'Rect', 'Whisker', 'Band', 'VLine', 'ErrorLine',
-           'ErrorCircle', 'BoxWhisker', 'get_glyph', 'get_glyph_list']
+__all__ = [
+    "Line",
+    "Circle",
+    "Diamond",
+    "Square",
+    "Triangle",
+    "Ray",
+    "HBar",
+    "VBar",
+    "Rect",
+    "Whisker",
+    "Band",
+    "VLine",
+    "ErrorLine",
+    "ErrorCircle",
+    "BoxWhisker",
+    "get_glyph",
+    "get_glyph_list",
+]
 
 
 # -- Glyphs -- #
 class BaseGlyph(object):
     """ Base class for glyphs. """
+
     method = None
     default_kwargs = MappingProxyType({})
 
-    def __init__(self, x_arg='x', y_arg='y', **kwargs):
+    def __init__(self, x_arg="x", y_arg="y", **kwargs):
         """ Constructor.
 
         Parameters
@@ -37,41 +55,47 @@ class BaseGlyph(object):
 
 class Line(BaseGlyph):
     """ A line glyph. """
-    method = 'line'
+
+    method = "line"
 
 
 class Circle(BaseGlyph):
     """ A circle glyph. """
-    method = 'circle'
+
+    method = "circle"
 
 
 class Diamond(BaseGlyph):
     """ A diamond glyph. """
-    method = 'diamond'
+
+    method = "diamond"
 
 
 class Square(BaseGlyph):
     """ A square glyph. """
-    method = 'square'
+
+    method = "square"
 
 
 class Triangle(BaseGlyph):
     """ A triangle glyph. """
-    method = 'triangle'
+
+    method = "triangle"
 
 
 class Ray(BaseGlyph):
     """ A ray glyph. """
-    method = 'ray'
-    default_kwargs = MappingProxyType({'length': 0, 'angle': 0})
+
+    method = "ray"
+    default_kwargs = MappingProxyType({"length": 0, "angle": 0})
 
 
 class HBar(BaseGlyph):
     """ A horizontal bar glyph. """
 
-    method = 'hbar'
+    method = "hbar"
 
-    def __init__(self, height, x_arg='right', y_arg='y', other=0., **kwargs):
+    def __init__(self, height, x_arg="right", y_arg="y", other=0.0, **kwargs):
         """ Constructor.
 
         Parameters
@@ -94,24 +118,23 @@ class HBar(BaseGlyph):
             Additional keyword arguments to be passed to the underlying
             bokeh glyph(s).
         """
-        if x_arg == 'left':
-            if 'right' not in kwargs:
-                kwargs.update({'right': other})
-        elif x_arg == 'right':
-            if 'left' not in kwargs:
-                kwargs.update({'left': other})
+        if x_arg == "left":
+            if "right" not in kwargs:
+                kwargs.update({"right": other})
+        elif x_arg == "right":
+            if "left" not in kwargs:
+                kwargs.update({"left": other})
         else:
-            raise ValueError('Unrecognized x_arg')
-        super(HBar, self).__init__(
-            x_arg, y_arg, height=height, **kwargs)
+            raise ValueError("Unrecognized x_arg")
+        super(HBar, self).__init__(x_arg, y_arg, height=height, **kwargs)
 
 
 class VBar(BaseGlyph):
     """ A vertical bar glyph. """
 
-    method = 'vbar'
+    method = "vbar"
 
-    def __init__(self, width, x_arg='x', y_arg='top', other=0., **kwargs):
+    def __init__(self, width, x_arg="x", y_arg="top", other=0.0, **kwargs):
         """ Constructor.
 
         Parameters
@@ -134,22 +157,23 @@ class VBar(BaseGlyph):
             Additional keyword arguments to be passed to the underlying
             bokeh glyph(s).
         """
-        if y_arg == 'top':
-            if 'bottom' not in kwargs:
-                kwargs.update({'bottom': other})
-        elif y_arg == 'bottom':
-            if 'top' not in kwargs:
-                kwargs.update({'top': other})
+        if y_arg == "top":
+            if "bottom" not in kwargs:
+                kwargs.update({"bottom": other})
+        elif y_arg == "bottom":
+            if "top" not in kwargs:
+                kwargs.update({"top": other})
         else:
-            raise ValueError('Unrecognized x_arg')
+            raise ValueError("Unrecognized x_arg")
         super(VBar, self).__init__(x_arg, y_arg, width=width, **kwargs)
 
 
 class Rect(BaseGlyph):
     """ A rectangle glyph. """
-    method = 'rect'
 
-    def __init__(self, width, height, x_arg='x', y_arg='y', **kwargs):
+    method = "rect"
+
+    def __init__(self, width, height, x_arg="x", y_arg="y", **kwargs):
         """ Constructor.
 
         Parameters
@@ -173,7 +197,8 @@ class Rect(BaseGlyph):
             bokeh glyph(s).
         """
         super(Rect, self).__init__(
-            x_arg, y_arg, width=width, height=height, **kwargs)
+            x_arg, y_arg, width=width, height=height, **kwargs
+        )
 
 
 # -- Compat Glyphs -- #
@@ -186,7 +211,7 @@ class Whisker(BaseGlyphCompat):
 
     method = _Whisker
 
-    def __init__(self, x_arg='base', y_arg='upper', other=0., **kwargs):
+    def __init__(self, x_arg="base", y_arg="upper", other=0.0, **kwargs):
         """ Constructor.
 
         Parameters
@@ -206,14 +231,14 @@ class Whisker(BaseGlyphCompat):
             Additional keyword arguments to be passed to the underlying
             bokeh glyph(s).
         """
-        if y_arg == 'upper':
-            if 'lower' not in kwargs:
-                kwargs.update({'lower': other})
-        elif y_arg == 'lower':
-            if 'upper' not in kwargs:
-                kwargs.update({'upper': other})
+        if y_arg == "upper":
+            if "lower" not in kwargs:
+                kwargs.update({"lower": other})
+        elif y_arg == "lower":
+            if "upper" not in kwargs:
+                kwargs.update({"upper": other})
         else:
-            raise ValueError('Unrecognized y_arg')
+            raise ValueError("Unrecognized y_arg")
         super(Whisker, self).__init__(x_arg, y_arg, **kwargs)
 
 
@@ -222,7 +247,7 @@ class Band(BaseGlyphCompat):
 
     method = _Band
 
-    def __init__(self, x_arg='base', y_arg='upper', other=0., **kwargs):
+    def __init__(self, x_arg="base", y_arg="upper", other=0.0, **kwargs):
         """ Constructor.
 
         Parameters
@@ -242,14 +267,14 @@ class Band(BaseGlyphCompat):
             Additional keyword arguments to be passed to the underlying
             bokeh glyph(s).
         """
-        if y_arg == 'upper':
-            if 'lower' not in kwargs:
-                kwargs.update({'lower': other})
-        elif y_arg == 'lower':
-            if 'upper' not in kwargs:
-                kwargs.update({'upper': other})
+        if y_arg == "upper":
+            if "lower" not in kwargs:
+                kwargs.update({"lower": other})
+        elif y_arg == "lower":
+            if "upper" not in kwargs:
+                kwargs.update({"upper": other})
         else:
-            raise ValueError('Unrecognized y_arg')
+            raise ValueError("Unrecognized y_arg")
         super(Band, self).__init__(x_arg, y_arg, **kwargs)
 
 
@@ -259,7 +284,7 @@ class CompositeGlyph(object):
 
     default_kwargs = MappingProxyType({})
 
-    def __init__(self, glyphs, x_arg='x', y_arg='y', **kwargs):
+    def __init__(self, glyphs, x_arg="x", y_arg="y", **kwargs):
         """ Constructor.
 
         Parameters
@@ -276,8 +301,9 @@ class CompositeGlyph(object):
         """
         glyph_kwargs = dict(self.default_kwargs)
         glyph_kwargs.update(kwargs)
-        self.glyphs = [g(x_arg=x_arg, y_arg=y_arg, **glyph_kwargs)
-                       for g in glyphs]
+        self.glyphs = [
+            g(x_arg=x_arg, y_arg=y_arg, **glyph_kwargs) for g in glyphs
+        ]
 
     def __iter__(self):
         return iter(self.glyphs)
@@ -289,18 +315,22 @@ class CompositeGlyph(object):
 class VLine(CompositeGlyph):
     """ A collection of vertical lines. """
 
-    default_kwargs = MappingProxyType({
-        'length': 0,
-        'line_width': 1,
-        'angle_units': 'deg',
-        'color': 'grey',
-        'alpha': 0.5})
+    default_kwargs = MappingProxyType(
+        {
+            "length": 0,
+            "line_width": 1,
+            "angle_units": "deg",
+            "color": "grey",
+            "alpha": 0.5,
+        }
+    )
 
-    def __init__(self, x_arg='x', y_arg='y', **kwargs):
+    def __init__(self, x_arg="x", y_arg="y", **kwargs):
         super(VLine, self).__init__(
-            [Ray, Ray], x_arg=x_arg, y_arg=y_arg, **kwargs)
-        self.glyphs[0].glyph_kwargs['angle'] = 90
-        self.glyphs[1].glyph_kwargs['angle'] = 270
+            [Ray, Ray], x_arg=x_arg, y_arg=y_arg, **kwargs
+        )
+        self.glyphs[0].glyph_kwargs["angle"] = 90
+        self.glyphs[1].glyph_kwargs["angle"] = 270
 
 
 class ErrorLine(CompositeGlyph):
@@ -324,10 +354,9 @@ class ErrorLine(CompositeGlyph):
             bokeh glyph(s).
         """
         self.glyphs = [Line(**kwargs)]
-        kwargs.pop('color', None)
-        kwargs.pop('legend', None)
-        self.glyphs.append(
-            Whisker(lower=lower, upper=upper, **kwargs))
+        kwargs.pop("color", None)
+        kwargs.pop("legend", None)
+        self.glyphs.append(Whisker(lower=lower, upper=upper, **kwargs))
 
 
 class ErrorCircle(CompositeGlyph):
@@ -351,16 +380,15 @@ class ErrorCircle(CompositeGlyph):
             bokeh glyph(s).
         """
         self.glyphs = [Circle(**kwargs)]
-        kwargs.pop('color', None)
-        kwargs.pop('legend', None)
-        self.glyphs.append(
-            Whisker(lower=lower, upper=upper, **kwargs))
+        kwargs.pop("color", None)
+        kwargs.pop("legend", None)
+        self.glyphs.append(Whisker(lower=lower, upper=upper, **kwargs))
 
 
 class BoxWhisker(CompositeGlyph):
     """ A box-whisker glyph. """
 
-    default_kwargs = MappingProxyType({'line_color': 'black'})
+    default_kwargs = MappingProxyType({"line_color": "black"})
 
     def __init__(self, width, q_lower, w_lower, q_upper, w_upper, **kwargs):
         """ Constructor.
@@ -395,11 +423,13 @@ class BoxWhisker(CompositeGlyph):
         glyph_kwargs.update(kwargs)
         self.glyphs = [
             VBar(width, bottom=q_lower, **glyph_kwargs),
-            VBar(width, y_arg='bottom', top=q_upper, **glyph_kwargs)]
-        glyph_kwargs.pop('color', None)
-        glyph_kwargs.pop('legend', None)
+            VBar(width, y_arg="bottom", top=q_upper, **glyph_kwargs),
+        ]
+        glyph_kwargs.pop("color", None)
+        glyph_kwargs.pop("legend", None)
         self.glyphs.append(
-            Whisker(lower=w_lower, upper=w_upper, **glyph_kwargs))
+            Whisker(lower=w_lower, upper=w_upper, **glyph_kwargs)
+        )
 
 
 def get_glyph(name, *args, **kwargs):
@@ -415,26 +445,28 @@ def get_glyph(name, *args, **kwargs):
     glyph : BaseGlyph
         An instance of the corresponding glyph class.
     """
-    glyphs = {'line': Line,
-              'circle': Circle,
-              'diamond': Diamond,
-              'square': Square,
-              'triangle': Triangle,
-              'ray': Ray,
-              'hbar': HBar,
-              'vbar': VBar,
-              'rect': Rect,
-              'vline': VLine,
-              'whisker': Whisker,
-              'band': Band,
-              'errorline': ErrorLine,
-              'errorcircle': ErrorCircle,
-              'boxwhisker': BoxWhisker}
+    glyphs = {
+        "line": Line,
+        "circle": Circle,
+        "diamond": Diamond,
+        "square": Square,
+        "triangle": Triangle,
+        "ray": Ray,
+        "hbar": HBar,
+        "vbar": VBar,
+        "rect": Rect,
+        "vline": VLine,
+        "whisker": Whisker,
+        "band": Band,
+        "errorline": ErrorLine,
+        "errorcircle": ErrorCircle,
+        "boxwhisker": BoxWhisker,
+    }
 
     try:
         return glyphs[name.lower()](*args, **kwargs)
     except KeyError:
-        raise ValueError('Unrecognized or unsupported glyph: ' + name)
+        raise ValueError("Unrecognized or unsupported glyph: " + name)
 
 
 def get_glyph_list(glyphs):
@@ -448,6 +480,7 @@ def get_glyph_list(glyphs):
             glyphs = [glyphs]
         else:
             glyphs = [
-                get_glyph(g) if isinstance(g, str) else g for g in glyphs]
+                get_glyph(g) if isinstance(g, str) else g for g in glyphs
+            ]
 
     return glyphs