+++ xrview/notebook/utils.py
@@ -1,19 +1,11 @@
 """ ``xrview.utils`` """
 import json
 import re
-from six.moves.urllib.parse import urljoin
+from urllib.parse import urljoin
 
 import ipykernel
 import requests
-
-try:  # Python 3
-    from notebook.notebookapp import list_running_servers
-except ImportError:  # Python 2
-    import warnings
-    from IPython.utils.shimmodule import ShimWarning
-    with warnings.catch_warnings():
-        warnings.simplefilter("ignore", category=ShimWarning)
-        from IPython.html.notebookapp import list_running_servers
+from notebook.notebookapp import list_running_servers
 
 
 def get_kernel_id():
@@ -26,7 +18,8 @@ def get_kernel_id():
     """
 
     return re.search(
-        'kernel-(.*).json', ipykernel.connect.get_connection_file()).group(1)
+        "kernel-(.*).json", ipykernel.connect.get_connection_file()
+    ).group(1)
 
 
 def get_notebook_url():
@@ -44,10 +37,12 @@ def get_notebook_url():
 
     for ss in servers:
         try:
-            response = requests.get(urljoin(ss['url'], 'api/sessions'),
-                                    params={'token': ss.get('token', '')})
+            response = requests.get(
+                urljoin(ss["url"], "api/sessions"),
+                params={"token": ss.get("token", "")},
+            )
             for nn in json.loads(response.text):
-                if nn['kernel']['id'] == kernel_id:
-                    return ss['url'][:-1]
+                if nn["kernel"]["id"] == kernel_id:
+                    return ss["url"][:-1]
         except Exception:
             pass