+++ HISTORY.rst
@@ -2,7 +2,16 @@
 History
 =======
 
-0.1.0 (tbd)
-------------------
+0.2.0 (Sep 8, 2020)
+-------------------
 
-* First release on PyPI.
+Breaking changes
+~~~~~~~~~~~~~~~~
+
+* Dropped compatibility with Python 2.7 and 3.5 and bokeh versions < 2.2
+
+0.1.0 (Sep 8, 2020)
+-------------------
+
+* First official release
+* Last release compatible with Python 2.7/3.5